Fix MIDI out dropping notes when playing from sequencer
Needed to buffer notes from the sequencer, they were coming quickly enough that they were overwriting the single note buffer previously used.
That was good enough for local playing, but having several notes firing at the same time would cause all but the last one to be overwritten, and so not sent.